Reservation && Guest DataBase

+++ src/main/java/pl/chylu/domain/reservation/Reservation.java
@@ -8,13 +8,13 @@ import java.time.LocalDateTime;
 
 public class Reservation {
 
-    private final int id;
+    private final long id;
     private final Room room;
     private final Guest guest;
     private final LocalDateTime from;
     private final LocalDateTime to;
 
-    public Reservation(int id, Room room, Guest guest, LocalDateTime from, LocalDateTime to) {
+    public Reservation(long id, Room room, Guest guest, LocalDateTime from, LocalDateTime to) {
         this.id = id;
         this.room = room;
         this.guest = guest;
@@ -32,7 +32,7 @@ public class Reservation {
                 System.getProperty("line.separator"));
     }
 
-    public int getId() {
+    public long getId() {
         return this.id;
     }
