+++ src/main/java/pl/chylu/domain/room/RoomRepository.java
@@ -3,11 +3,18 @@ package pl.chylu.domain.room;
 import java.util.List;
 
 public interface RoomRepository {
+
     void saveAll();
+
     void readAll();
-    void remove(int id);
-    void edit(int id, int number, List<BedType> bedTypes);
-    Room getById(int id);
+
+    void remove(long id);
+
+    void edit(long id, int number, List<BedType> bedTypes);
+
+    Room getById(long id);
+
     Room createNewRoom(int number, List<BedType> bedTypes);
+
     List<Room> getAllRooms();
 }
